Initial support for Vagrant
During development it's not always possible or convenient to
have access to an OpenStack cloud to host the virtual machines
for the infrastructure.

This change adds a Vagrantfile which utilizes libvirt as the
provisioning backend for deploying the base infrastructure
as managed via Vagrant.

These changes have been tested against both a Vagrant deployment with
libvirt, and against the existing OpenStack deployment system.