Add /var/log/artib directory creation
That path is hardcoded in the role we use for building images.
The role tries to create it, but as we run with stack user, is
not capable to do it as it's outside /home/stack directory.
A proper fix needs to land on that role, but as it's on the process
of being moved from git to openstack, it can take long. As a
workaround, create the directory on our role.