Fix some small issues that was causing a deploy failure
Added a new variable to vars/main.yml so that there was a default
value set. Also changed from using the yum Ansible module to the
package module so that we can support other virthosts (like Fedora).

Added the installation of python-firewall package to resolve issues
when executing a block with firewall commands. Found inadvertantly.