Fix FileBeat installation issues
This fix changes the 'filebeat version' to be 5.x.

The upstream repository path changed and now access to the
1.x versions seems to be gone. I don't love this fix but
it'll allow things to pass and install. I haven't done any
testing with FileBeat in a bit, so will need to revisit.

I plan to add changes to upstream torian.filebeat ansible
role so that I can lock to a specific FileBeat version in
the 5.x branch.

Also fix a missing var_files include.